Device and repr fixes
Updated device semantics of CompositeSpec to match that of TensorDict,
i.e. it is possible to have device=None and entries on different
devices.

Created new mixin class that doesn't inherit from TensorSpec. This is
necessary to ensure that __torch_function__ dispatches to the right
method when stacking LazyStackedCompositeSpec

repr now works correctly for all spec types.