fix buffer __iter__ for samplers without replacement + prefetch
For replay buffers with prefetch enabled while having a sampler without replacement (that can "run out"), the buffer's `__iter__` method will not yield all samples. This is because its while loop terminates on sampler `ran_out` which is set as soon as a prefetch job is executed that triggers running out. However, there might still be prefetch jobs in the queue that could yield samples. Therefore, new flags are added for adding jobs to the prefetch queue and for terminating the yield loop only if available prefetch jobs are completed